+++ DodgerGame/src/main/java/Model/GameWorld.java
@@ -3,7 +3,6 @@ package Model;
 import Model.Entities.HitBox;
 
 import Model.Entities.Player.Spaceship;
-import Model.Entities.Player.SpaceshipFactory;
 import Model.Entities.AbstractGameObject;
 
 import java.util.ArrayList;
