+++ DodgerGame/src/main/java/View/Menu/HighScoreMenu.java
@@ -1,6 +1,6 @@
 package View.Menu;
 
-import Model.Handlers.HighScoreHandler;
+import Model.HighScoreHandler;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.paint.Color;
