+++ DodgerGame/src/test/java/HighScoreHandlerTest.java
@@ -1,5 +1,5 @@
 
-import Model.Handlers.HighScoreHandler;
+import Model.HighScoreHandler;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
